[COLAB-2961] rename flagging-client to moderation-client
