--- microservices/clients/flagging-client/pom.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.xcolab</groupId>
-        <artifactId>microservice-clients</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>flagging-client</artifactId>
-    <packaging>jar</packaging>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-starter-feign</artifactId>
-            <version>${spring-cloud.feign}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.xcolab</groupId>
-            <artifactId>xcolab-utils</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.xcolab</groupId>
-            <artifactId>members-client</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.xcolab</groupId>
-            <artifactId>pojo-generator-maven-plugin</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.xcolab</groupId>
-                <artifactId>pojo-generator-maven-plugin</artifactId>
-                <version>${project.version}</version>
-                <executions>
-                    <execution>
-                        <phase>generate-sources</phase>
-                        <goals>
-                            <goal>generatePojos</goal>
-                        </goals>
-                    </execution>
-                </executions>
-                <configuration>
-                    <interfaceDirectory>src/main/java/org/xcolab/client/flagging/pojo</interfaceDirectory>
-                    <implementationPackageSuffix>tables.pojos</implementationPackageSuffix>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
