--- microservices/clients/flagging-client/src/main/java/org/xcolab/client/flagging/exceptions/ReportNotFoundException.java
@@ -1,8 +1,0 @@
-package org.xcolab.client.flagging.exceptions;
-
-public class ReportNotFoundException extends Exception {
-
-    public ReportNotFoundException(long reportId) {
-        super("Report with id " + reportId + " not found");
-    }
-}
