+++ microservices/clients/moderation-client/src/main/java/org/xcolab/client/moderation/exceptions/ReportNotFoundException.java
@@ -1,0 +1,8 @@
+package org.xcolab.client.moderation.exceptions;
+
+public class ReportNotFoundException extends Exception {
+
+    public ReportNotFoundException(long reportId) {
+        super("Report with id " + reportId + " not found");
+    }
+}
