+++ microservices/services/moderation-service/pom.xml
@@ -22,7 +22,7 @@
         </dependency>
         <dependency>
             <groupId>org.xcolab</groupId>
-            <artifactId>flagging-client</artifactId>
+            <artifactId>moderation-client</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
