+++ microservices/services/moderation-service/src/main/java/org/xcolab/service/moderation/ModerationServiceApplication.java
@@ -1,0 +1,15 @@
+package org.xcolab.service.moderation;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class ModerationServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ModerationServiceApplication.class, args);
+    }
+
+}
