+++ microservices/services/moderation-service/src/main/java/org/xcolab/service/moderation/exceptions/NotFoundException.java
@@ -1,0 +1,15 @@
+package org.xcolab.service.moderation.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value= HttpStatus.NOT_FOUND, reason="Object Not Found")
+public class NotFoundException extends Exception {
+
+    public NotFoundException() {
+    }
+
+    public NotFoundException(String msg) {
+        super(msg);
+    }
+}
