--- util/xcolab-utils/src/main/java/org/xcolab/util/enums/flagging/ManagerAction.java
@@ -1,5 +1,0 @@
-package org.xcolab.util.enums.flagging;
-
-public enum ManagerAction {
-    PENDING, APPROVE, REVERT, REMOVE
-}
