--- util/xcolab-utils/src/main/java/org/xcolab/util/enums/flagging/TargetType.java
@@ -1,5 +1,0 @@
-package org.xcolab.util.enums.flagging;
-
-public enum TargetType {
-    COMMENT, PROPOSAL
-}
