+++ util/xcolab-utils/src/main/java/org/xcolab/util/enums/moderation/ManagerAction.java
@@ -1,0 +1,5 @@
+package org.xcolab.util.enums.moderation;
+
+public enum ManagerAction {
+    PENDING, APPROVE, REVERT, REMOVE
+}
