+++ view/src/main/java/org/xcolab/view/config/spring/StaticInjector.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Component;
 
 import org.xcolab.client.content.IContentClient;
 import org.xcolab.client.content.IFileClient;
-import org.xcolab.client.flagging.IFlaggingClient;
+import org.xcolab.client.moderation.IModerationClient;
 import org.xcolab.client.search.ISearchClient;
-import org.xcolab.view.moderation.FlaggingController;
+import org.xcolab.view.moderation.ModerationController;
 import org.xcolab.view.pages.contestmanagement.controller.manager.FlaggingTabController;
-import org.xcolab.view.pages.contestmanagement.wrappers.FlaggingReportWrapper;
+import org.xcolab.view.pages.contestmanagement.wrappers.ModerationReportWrapper;
 import org.xcolab.view.pages.loginregister.ImageUploadUtils;
 import org.xcolab.view.pages.proposals.view.proposal.tabs.ProposalDescriptionTabController;
 import org.xcolab.view.pages.search.paging.SearchDataPage;
@@ -32,7 +32,7 @@ public class StaticInjector implements ApplicationRunner {
     private IContentClient contentClient;
 
     @Autowired
-    private IFlaggingClient flaggingClient;
+    private IModerationClient moderationClient;
 
     @Override
     public void run(ApplicationArguments applicationArguments) throws Exception {
@@ -40,11 +40,11 @@ public class StaticInjector implements ApplicationRunner {
         ImageUploadUtils.setFileClient(fileClient);
         LoadContentArticleTag.setContentClient(contentClient);
 
-        FlaggingTabController.setFlaggingClient(flaggingClient);
-        FlaggingController.setFlaggingClient(flaggingClient);
-        LoadThreadStartTag.setFlaggingClient(flaggingClient);
-        FlaggingReportWrapper.setFlaggingClient(flaggingClient);
-        ProposalDescriptionTabController.setFlaggingClient(flaggingClient);
-        DiscussionPermissions.setFlaggingClient(flaggingClient);
+        FlaggingTabController.setmoderationClient(moderationClient);
+        ModerationController.setmoderationClient(moderationClient);
+        LoadThreadStartTag.setmoderationClient(moderationClient);
+        ModerationReportWrapper.setmoderationClient(moderationClient);
+        ProposalDescriptionTabController.setmoderationClient(moderationClient);
+        DiscussionPermissions.setmoderationClient(moderationClient);
     }
 }
