--- view/src/main/java/org/xcolab/view/moderation/FlaggingController.java
@@ -1,51 +1,0 @@
-package org.xcolab.view.moderation;
-
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import org.xcolab.client.flagging.IFlaggingClient;
-import org.xcolab.client.members.pojo.Member;
-import org.xcolab.util.enums.flagging.TargetType;
-
-import java.io.UnsupportedEncodingException;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@RestController
-public class FlaggingController {
-
-    private static IFlaggingClient flaggingClient;
-
-    public static void setFlaggingClient(IFlaggingClient flaggingClient) {
-        FlaggingController.flaggingClient = flaggingClient;
-    }
-
-    @PostMapping("/flagging/report")
-    public ResponseJson report(HttpServletRequest request, HttpServletResponse response,
-            @RequestParam TargetType targetType, @RequestParam long targetAdditionalId,
-            @RequestParam String reason, @RequestParam String comment, @RequestParam long targetId,
-            Member loggedInMember) throws UnsupportedEncodingException {
-
-        request.setCharacterEncoding("UTF-8");
-
-        flaggingClient
-                .report(loggedInMember, targetId, targetAdditionalId, targetType, reason, comment);
-
-        return new ResponseJson(true);
-    }
-
-    private static class ResponseJson {
-
-        private final boolean success;
-
-        private ResponseJson(boolean success) {
-            this.success = success;
-        }
-
-        public boolean isSuccess() {
-            return success;
-        }
-    }
-}
