+++ view/src/main/java/org/xcolab/view/pages/contestmanagement/entities/ContestManagerTabs.java
@@ -24,7 +24,7 @@ public enum ContestManagerTabs implements TabEnum {
     TEAMS("Teams", "team",
             TabPermissionAlgorithm.adminOnlyViewAndEdit,
             TabActivityCountAlgorithm.alwaysZero),
-    FLAGGING("Flagging", "flagging",
+    FLAGGING("Flagging", "moderation",
             TabPermissionAlgorithm.adminOnlyViewAndEdit,
             TabActivityCountAlgorithm.alwaysZero),
     COLLECTION_CARDS("Collection Cards", "collectionCard",
