+++ view/src/main/java/org/xcolab/view/taglibs/xcolab/jspTags/discussion/LoadThreadStartTag.java
@@ -11,11 +11,10 @@ import org.xcolab.client.comment.exceptions.ThreadNotFoundException;
 import org.xcolab.client.comment.pojo.Category;
 import org.xcolab.client.comment.pojo.CategoryGroup;
 import org.xcolab.client.comment.pojo.CommentThread;
-import org.xcolab.client.flagging.IFlaggingClient;
-import org.xcolab.client.flagging.pojo.IReportTarget;
+import org.xcolab.client.moderation.IModerationClient;
+import org.xcolab.client.moderation.pojo.IReportTarget;
 import org.xcolab.commons.exceptions.ReferenceResolutionException;
-import org.xcolab.util.enums.flagging.TargetType;
-import org.xcolab.view.pages.contestmanagement.controller.manager.FlaggingTabController;
+import org.xcolab.util.enums.moderation.TargetType;
 import org.xcolab.view.taglibs.xcolab.jspTags.discussion.wrappers.NewMessageWrapper;
 
 import java.util.List;
@@ -27,10 +26,10 @@ public class LoadThreadStartTag extends BodyTagSupport {
 
     private static final Logger _log = LoggerFactory.getLogger(LoadThreadStartTag.class);
 
-    private static IFlaggingClient flaggingClient;
+    private static IModerationClient moderationClient;
 
-    public static void setFlaggingClient(IFlaggingClient flaggingClient) {
-        LoadThreadStartTag.flaggingClient = flaggingClient;
+    public static void setmoderationClient(IModerationClient moderationClient) {
+        LoadThreadStartTag.moderationClient = moderationClient;
     }
     private long threadId;
     private long categoryId;
@@ -83,7 +82,7 @@ public class LoadThreadStartTag extends BodyTagSupport {
             pageContext.setAttribute("newMessage", new NewMessageWrapper());
             pageContext.setAttribute("discussionPermissions", discussionPermissions);
             final List<IReportTarget> reportTargets =
-                    flaggingClient.listReportTargets(TargetType.COMMENT);
+                    moderationClient.listReportTargets(TargetType.COMMENT);
             pageContext.setAttribute("reportTargets", reportTargets);
 
         } catch (ThreadNotFoundException e) {
