+++ view/src/main/webapp/WEB-INF/tags/discussionComments.tagx
@@ -26,7 +26,7 @@
         <!--@elvariable id="newMessage" type="org.xcolab.view.taglibs.xcolab.jspTags.discussion.wrappers.NewMessageWrapper"-->
         <!--@elvariable id="discussionPermissions" type="org.xcolab.view.taglibs.xcolab.jspTags.discussion.DiscussionPermissions"-->
         <!--@elvariable id="shareTitle" type="java.lang.String"-->
-        <!--@elvariable id="reportTargets" type="java.util.List<org.xcolab.client.flagging.pojo.ReportTarget>"-->
+        <!--@elvariable id="reportTargets" type="java.util.List<org.xcolab.client.moderation.pojo.ReportTarget>"-->
         <!--@elvariable id="ontologySpaces" type="java.util.List<org.xcolab.client.contest.pojo.ontology.OntologySpace>"-->
 
 		<xcolab:requireLibrary name="CKEditor" />
